+++ R/db-sql.R
@@ -452,6 +452,7 @@ sql_query_multi_join <- function(
   by_list,
   select,
   ...,
+  distinct = FALSE,
   lvl = 0
 ) {
   check_dots_used()
@@ -496,6 +497,7 @@ sql_query_multi_join.DBIConnection <- function(
   by_list,
   select,
   ...,
+  distinct = FALSE,
   lvl = 0
 ) {
   if (vctrs::vec_duplicate_any(table_names)) {
@@ -520,7 +522,7 @@ sql_query_multi_join.DBIConnection <- function(
   }
 
   clauses <- list2(
-    sql_clause_select(con, select),
+    sql_clause_select(con, select, distinct),
     sql_clause_from(from),
     !!!out
   )