+++ NEWS.md
@@ -1,5 +1,8 @@
 # dbplyr (development version)
 
+* `summarise()` now reports grouping immediately, rather than when you summarise.
+* `sql_optimise()` has been removed. It was only used for two cases (filter + summarise and arrange + summarise), and these are now handled at a higher level (#1720).
+* `distinct()` after a join no longer creates a subquery (#722).
 * New `bind_queries()` makes it easy to combine multiple lazy queries using   `UNION ALL` (#1342).
 * `slice_*()` now handles missing values in line with the documentation, i.e. they are always removed (#1599).
 * Internal testing functions `src_test()`, `test_frame()` and `test_load()`, `test_register_src()` and `test_register_con()` have been removed.